- motd.c, send.c: use %zu conversion specifier for size_t types
git-svn-id: svn://svn.ircd-hybrid.org/svnroot/ircd-hybrid/branches/8.2.x@6536 82007160-df01-0410-b94d-b575c5fd34c7

+++ src/send.c
@@ -90,9 +90,9 @@ send_message(struct Client *to, struct dbuf_block *buf)
   {
     if (IsServer(to))
       sendto_realops_flags(UMODE_SERVNOTICE, L_ALL, SEND_NOTICE,
-                           "Max SendQ limit exceeded for %s: %lu > %u",
+                           "Max SendQ limit exceeded for %s: %zu > %u",
                            get_client_name(to, HIDE_IP),
-                           (unsigned long)(dbuf_length(&to->connection->buf_sendq) + buf->size),
+                           (dbuf_length(&to->connection->buf_sendq) + buf->size),
                            get_sendq(&to->connection->confs));
 
     if (IsClient(to))
